+++ programs/drift/src/instructions/pyth_pull_oracle.rs
@@ -159,7 +159,6 @@ pub fn handle_post_multi_pyth_pull_oracle_updates_atomic<'c: 'info, 'info>(
         let next_timestamp = get_timestamp_from_price_update_message(&merkle_price_update.message)?;
 
         drop(price_feed_account_data);
-        drop(price_feed_account);
 
         if next_timestamp > current_timestamp {
             pyth_solana_receiver_sdk::cpi::post_update_atomic(