+++ sdk/src/driftClient.ts
@@ -7512,6 +7512,7 @@ export class DriftClient {
 
 		const [pullIx, _responses, success] = await feedAccount.fetchUpdateIx({
 			numSignatures,
+			// @ts-ignore :: TODO someone needs to look at this
 			feedConfigs: this.sbProgramFeedConfigs.get(feed.toString()),
 		});
 		if (!success) {