+++ sdk/src/math/margin.ts
@@ -5,6 +5,8 @@ import {
 	ZERO,
 	BID_ASK_SPREAD_PRECISION,
 	AMM_RESERVE_PRECISION,
+	MAX_PREDICTION_PRICE,
+	BASE_PRECISION,
 } from '../constants/numericConstants';
 import { BN } from '@coral-xyz/anchor';
 import { OraclePriceData } from '../oracles/types';
@@ -99,6 +101,14 @@ export function calculateOraclePriceForPerpMargin(
 	return marginPrice;
 }
 
+/**
+ * This is _not_ the same as liability value as for prediction markets, the liability for the short in prediction market is (1 - oracle price) * base
+ * See {@link calculatePerpLiabilityValue} to get the liabiltiy value
+ * @param market
+ * @param perpPosition
+ * @param oraclePriceData
+ * @param includeOpenOrders
+ */
 export function calculateBaseAssetValueWithOracle(
 	market: PerpMarketAccount,
 	perpPosition: PerpPosition,
@@ -111,21 +121,76 @@ export function calculateBaseAssetValueWithOracle(
 	}
 
 	const baseAssetAmount = includeOpenOrders
-		? calculateWorstCaseBaseAssetAmount(perpPosition)
+		? calculateWorstCaseBaseAssetAmount(
+				perpPosition,
+				market,
+				oraclePriceData.price
+		  )
 		: perpPosition.baseAssetAmount;
 
 	return baseAssetAmount.abs().mul(price).div(AMM_RESERVE_PRECISION);
 }
 
 export function calculateWorstCaseBaseAssetAmount(
-	perpPosition: PerpPosition
+	perpPosition: PerpPosition,
+	perpMarket: PerpMarketAccount,
+	oraclePrice: BN
 ): BN {
+	return calculateWorstCasePerpLiabilityValue(
+		perpPosition,
+		perpMarket,
+		oraclePrice
+	).worstCaseBaseAssetAmount;
+}
+
+export function calculateWorstCasePerpLiabilityValue(
+	perpPosition: PerpPosition,
+	perpMarket: PerpMarketAccount,
+	oraclePrice: BN
+): { worstCaseBaseAssetAmount: BN; worstCaseLiabilityValue: BN } {
 	const allBids = perpPosition.baseAssetAmount.add(perpPosition.openBids);
 	const allAsks = perpPosition.baseAssetAmount.add(perpPosition.openAsks);
 
-	if (allBids.abs().gt(allAsks.abs())) {
-		return allBids;
+	const isPredictionMarket = isVariant(perpMarket.contractType, 'prediction');
+	const allBidsLiabilityValue = calculatePerpLiabilityValue(
+		allBids,
+		oraclePrice,
+		isPredictionMarket
+	);
+	const allAsksLiabilityValue = calculatePerpLiabilityValue(
+		allAsks,
+		oraclePrice,
+		isPredictionMarket
+	);
+
+	if (allAsksLiabilityValue.gte(allBidsLiabilityValue)) {
+		return {
+			worstCaseBaseAssetAmount: allAsks,
+			worstCaseLiabilityValue: allAsksLiabilityValue,
+		};
 	} else {
-		return allAsks;
+		return {
+			worstCaseBaseAssetAmount: allBids,
+			worstCaseLiabilityValue: allBidsLiabilityValue,
+		};
+	}
+}
+
+export function calculatePerpLiabilityValue(
+	baseAssetAmount: BN,
+	oraclePrice: BN,
+	isPredictionMarket: boolean
+): BN {
+	if (isPredictionMarket) {
+		if (baseAssetAmount.gt(ZERO)) {
+			return baseAssetAmount.mul(oraclePrice).div(BASE_PRECISION);
+		} else {
+			return baseAssetAmount
+				.abs()
+				.mul(MAX_PREDICTION_PRICE.sub(oraclePrice))
+				.div(BASE_PRECISION);
+		}
+	} else {
+		return baseAssetAmount.abs().mul(oraclePrice).div(BASE_PRECISION);
 	}
 }