+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: init prediction markets ([#1152](https://github.com/drift-labs/protocol-v2/pull/1152))
+
 ### Fixes
 
 - program: make updateUserQuoteAssetInsuranceStake permissionless ([#1187](https://github.com/drift-labs/protocol-v2/pull/1187))